Fix package restrictions and invalid @Unique
