+++ src/main/java/net/gensokyoreimagined/nitori/access/IMixinChunkMapAccess.java
@@ -12,11 +12,8 @@
 //
 // You should have received a copy of the GNU General Public License
 // along with this program.  If not, see <https://www.gnu.org/licenses/>.
-package net.gensokyoreimagined.nitori.core.access;
+package net.gensokyoreimagined.nitori.access;
 
-import com.destroystokyo.paper.util.maplist.EntityList;
-
-public interface IMixinChunkEntitySlicesAccess {
-    @SuppressWarnings("EmptyMethod")
-    EntityList getEntities();
+public interface IMixinChunkMapAccess {
+    void gensouHacks$runOnTrackerMainThread(final Runnable runnable);
 }
