+++ src/main/java/net/gensokyoreimagined/nitori/access/IMixinChunkMap_TrackedEntityAccess.java
@@ -12,7 +12,7 @@
 //
 // You should have received a copy of the GNU General Public License
 // along with this program.  If not, see <https://www.gnu.org/licenses/>.
-package net.gensokyoreimagined.nitori.core.access;
+package net.gensokyoreimagined.nitori.access;
 
 import com.destroystokyo.paper.util.misc.PooledLinkedHashSets;
 import net.minecraft.server.level.ServerPlayer;
