+++ src/main/java/net/gensokyoreimagined/nitori/access/IMixinIteratorSafeOrderedReferenceSetAccess.java
@@ -12,8 +12,9 @@
 //
 // You should have received a copy of the GNU General Public License
 // along with this program.  If not, see <https://www.gnu.org/licenses/>.
-package net.gensokyoreimagined.nitori.core.access;
+package net.gensokyoreimagined.nitori.access;
 
-public interface IMixinChunkMapAccess {
-    void gensouHacks$runOnTrackerMainThread(final Runnable runnable);
+public interface IMixinIteratorSafeOrderedReferenceSetAccess {
+    @SuppressWarnings("EmptyMethod")
+    int getListSize();
 }
