+++ src/main/java/net/gensokyoreimagined/nitori/core/ChunkMapMixin.java
@@ -20,8 +20,8 @@ import it.unimi.dsi.fastutil.ints.Int2ObjectMap;
 import it.unimi.dsi.fastutil.objects.ReferenceOpenHashSet;
 import it.unimi.dsi.fastutil.objects.ReferenceSets;
 //import net.gensokyoreimagined.nitori.plugins.NitoriConfig;
-import net.gensokyoreimagined.nitori.core.access.IMixinChunkMapAccess;
-import net.gensokyoreimagined.nitori.core.access.IMixinChunkMap_TrackedEntityAccess;
+import net.gensokyoreimagined.nitori.access.IMixinChunkMapAccess;
+import net.gensokyoreimagined.nitori.access.IMixinChunkMap_TrackedEntityAccess;
 import net.gensokyoreimagined.nitori.tracker.MultithreadedTracker;
 import net.minecraft.server.level.ChunkMap;
 import net.minecraft.server.level.ServerLevel;
