+++ src/main/java/net/gensokyoreimagined/nitori/tracker/MultithreadedTracker.java
@@ -21,9 +21,9 @@ package net.gensokyoreimagined.nitori.tracker;
 import com.google.common.util.concurrent.ThreadFactoryBuilder;
 import io.papermc.paper.util.maplist.IteratorSafeOrderedReferenceSet;
 import io.papermc.paper.world.ChunkEntitySlices;
-import net.gensokyoreimagined.nitori.core.access.IMixinChunkEntitySlicesAccess;
-import net.gensokyoreimagined.nitori.core.access.IMixinChunkMap_TrackedEntityAccess;
-import net.gensokyoreimagined.nitori.core.access.IMixinIteratorSafeOrderedReferenceSetAccess;
+import net.gensokyoreimagined.nitori.access.IMixinChunkEntitySlicesAccess;
+import net.gensokyoreimagined.nitori.access.IMixinChunkMap_TrackedEntityAccess;
+import net.gensokyoreimagined.nitori.access.IMixinIteratorSafeOrderedReferenceSetAccess;
 import net.minecraft.server.MinecraftServer;
 import net.minecraft.server.level.ChunkMap;
 import net.minecraft.world.entity.Entity;
